refactor: rename args.py to models.py
