+++ rekordbox_edit/api/edit.py
@@ -6,7 +6,7 @@ from pyrekordbox.db6 import DjmdContent
 from sqlalchemy import select
 
 from rekordbox_edit.api._utils import _track_from_content
-from rekordbox_edit.args import EditPlanArgs, Track
+from rekordbox_edit.models import EditPlanArgs, Track
 from rekordbox_edit.query import get_filtered_content
 
 logger = logging.getLogger(__name__)
