+++ rekordbox_edit/cli/search.py
@@ -11,7 +11,7 @@ from rekordbox_edit._click import (
     track_ids_argument,
 )
 from rekordbox_edit.api.search import search
-from rekordbox_edit.args import FilterArgs
+from rekordbox_edit.models import FilterArgs
 from rekordbox_edit.cli._utils import _handle_stdin, _print_ids
 from rekordbox_edit.display import print_track_info
 from rekordbox_edit.logger import get_debug_file_path, set_level
